# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaksi Pekerjaan API",
        "description": "API untuk aplikasi **Transaksi Pekerjaan** — pencatatan arus kas per kegiatan\nbeserta perhitungan bagi hasil.\n\n### Alur autentikasi\n1. `POST /api/auth/login` dengan email + password → dapat `access_token`.\n2. Kirim `Authorization: Bearer <access_token>` untuk semua endpoint lain.\n3. Opsional: `POST /api/auth/biometric/enable` (butuh password lagi) →\n   dapat `biometric_token` yang disimpan aplikasi di Android Keystore.\n   Login berikutnya cukup sidik jari lewat `POST /api/auth/biometric/login`.\n\n### Aturan angka\nSemua nilai uang adalah **rupiah bulat** (integer, tanpa desimal).\nPersentase dikirim sebagai angka desimal biasa, mis. `1.75` untuk 1,75%.",
        "contact": {
            "email": "bayuapriansah10@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/kegiatan/{id}/bahan-baku": {
            "get": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Rincian bahan baku satu kegiatan",
                "description": "Rincian bahan baku per item beserta totalnya. Total inilah yang menjadi porsi \"Bahan Baku\" pada Biaya Pelaksanaan Real -- angkanya tidak pernah diketik manual.",
                "operationId": "bahanBakuIndex",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/BahanBakuItem"
                                                    }
                                                },
                                                "total_bahan_baku": {
                                                    "type": "integer",
                                                    "example": 142400000
                                                },
                                                "total_upah": {
                                                    "type": "integer",
                                                    "example": 67000000
                                                },
                                                "total_pelaksanaan": {
                                                    "type": "integer",
                                                    "example": 209400000
                                                },
                                                "boleh_ubah": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Tambah item bahan baku",
                "description": "Menambah satu baris bahan baku. Subtotal dihitung server dari qty x harga_satuan, dan total keseluruhan langsung memperbarui Biaya Pelaksanaan Real beserta seluruh angka profit.",
                "operationId": "bahanBakuStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "qty",
                                    "harga_satuan"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Besi beton 12mm"
                                    },
                                    "satuan": {
                                        "type": "string",
                                        "example": "batang"
                                    },
                                    "qty": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "harga_satuan": {
                                        "type": "integer",
                                        "example": 145000
                                    },
                                    "tanggal_beli": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "no_struk": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "toko": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item ditambahkan"
                    },
                    "403": {
                        "description": "Tidak punya hak"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bahan-baku/{id}": {
            "put": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Ubah item bahan baku",
                "description": "Mengubah satu baris bahan baku. Subtotal dan seluruh angka taksasi dihitung ulang.",
                "operationId": "bahanBakuUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tersimpan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Hapus item bahan baku",
                "description": "Menghapus satu baris bahan baku (soft delete). Total dan angka profit dihitung ulang.",
                "operationId": "bahanBakuDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Terhapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BahanBakuItem": {
                "description": "Satu baris rincian bahan baku. Subtotal dihitung server dari qty x harga_satuan; klien tidak pernah mengirimkannya.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kegiatan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama": {
                        "type": "string",
                        "example": "Besi beton 12mm"
                    },
                    "satuan": {
                        "type": "string",
                        "example": "batang"
                    },
                    "qty": {
                        "description": "Boleh pecahan, mis. 4.5 m3",
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "qty_formatted": {
                        "type": "string",
                        "example": "250"
                    },
                    "harga_satuan": {
                        "type": "integer",
                        "example": 145000
                    },
                    "harga_satuan_formatted": {
                        "type": "string",
                        "example": "Rp145.000"
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 36250000
                    },
                    "subtotal_formatted": {
                        "type": "string",
                        "example": "Rp36.250.000"
                    },
                    "tanggal_beli": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "no_struk": {
                        "type": "string",
                        "example": "INV/BJ/1207",
                        "nullable": true
                    },
                    "toko": {
                        "type": "string",
                        "example": "TB Sumber Jaya",
                        "nullable": true
                    },
                    "keterangan": {
                        "type": "string",
                        "nullable": true
                    },
                    "urutan": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan access_token dari endpoint login (tanpa kata \"Bearer\").",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}